[client] move stuff to separate class.
